backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.6.4
pyiceberg==0.6.0
deltalake==0.17.0
pyarrow==15.0.2
boto3==1.34.84
fsspec==2024.3.1
s3fs==2024.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
duckdb==0.10.1
